decoupling IParserState
